Add Ausgabe renderer module
