+++ scripts/test.cjs
@@ -14,6 +14,7 @@ const { runTopsDtosTests } = require("./tests/topsDtos.test.cjs");
 const { runTopsCloseFlowTests } = require("./tests/topsCloseFlow.test.cjs");
 const { runProtokollRouterFallbackTests } = require("./tests/protokollRouterFallback.test.cjs");
 const { runProjektverwaltungModuleTests } = require("./tests/projektverwaltungModule.test.cjs");
+const { runAusgabeModuleTests } = require("./tests/ausgabeModule.test.cjs");
 
 let failed = false;
 
@@ -74,6 +75,7 @@ async function main() {
   await runTopsScreenIntegrationTests(run);
   await runProtokollRouterFallbackTests(run);
   await runProjektverwaltungModuleTests(run);
+  await runAusgabeModuleTests(run);
 
   if (failed) {
     process.exitCode = 1;
