+++ src/renderer/modules/ausgabe/README.md
@@ -1,0 +1,11 @@
+# Ausgabe / Drucken / E-Mail
+
+Dieses Modul kapselt die rendererseitigen Einstiege fuer Ausgabe, Drucken und Mail-Anstoss.
+
+Enthaltene Bausteine:
+- `PrintModal`
+- `sendMailPayload`
+
+Hinweis:
+- Die eigentliche PDF-, Druck- und Outlook-Technik bleibt im Main-Prozess.
+- Das ist kein Sidebar-Modul und kein Eintrag im Modulkatalog.
