+++ src/renderer/modules/ausgabe/index.js
@@ -1,0 +1,2 @@
+export { default as PrintModal } from "./PrintModal.js";
+export { sendMailPayload } from "./sendMailPayload.js";
